+++ app/src/main/java/ru/oliverhd/test_weather/retrofit/WeatherRetrofitImpl.kt
@@ -1,4 +1,4 @@
-package ru.oliverhd.test_weather.repository
+package ru.oliverhd.test_weather.retrofit
 
 import com.google.gson.GsonBuilder
 import okhttp3.Interceptor
